refactor: move more hooks to their own files
