+++ src/player-area.js
@@ -2,7 +2,8 @@ import React, { useRef, useState } from "react";
 import { PanZoom } from "react-easy-panzoom";
 import Referentiel from "referentiel";
 import useAsyncEffect from "@n1ru4l/use-async-effect";
-import { loadImage, useLongPress, getOptimalDimensions } from "./util";
+import { loadImage, getOptimalDimensions } from "./util";
+import { useLongPress } from "./hooks/use-long-press";
 import { ObjectLayer } from "./object-layer";
 import { Toolbar } from "./toolbar";
 import styled from "@emotion/styled/macro";
